+++ x/posts/client/cli/tx.go
@@ -50,40 +50,40 @@ func GetCmdCreatePost(cdc *codec.Codec) *cobra.Command {
 		Use:   "create [subspace] [message] [allows-comments]",
 		Short: "Create a new post",
 		Long: fmt.Sprintf(`
-				Create a new post, specifying the subspace, message and whether or not it will allow for comments.
-				Optional media attachments and polls are also supported.
-				If you want to add one or more medias attachments, you have to use the --media flag.
-				You need to specify both media's URI and mime-type in this order separeted by a comma
-				Usage examples:
+Create a new post specifying the subspace, message and whether or not it will allow for comments.
+Optional media attachments and polls are also supported. See the below sections to know how to include them.
 
-				- tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "Hello world!" true
-				- tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "A post with multiple medias" false \
-                  --media "https://example.com/media1,text/plain"
-				- tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "A post with multiple medias" false \
-                  --media "https://example.com/media1,text/plain"
-				  --media "https://example.com/media2,application/json"
- 				  --poll-details "question=Which dog do you prefer?,multiple-answers=false,allows-answer-edits=true,end-date=2020-01-01T15:00:00.000Z" \
- 				  --poll-answer "Beagle" \
-                  --poll-answer "Pug" \
-                  --poll-answer "German Sheperd"
-				
-				If you want to add a poll to your post you need to specify it through two flags:
-					1. --poll-details 
-                       where you specify a map of the following fields:
-						* The question of the poll (key: question)
-						* The end date of your poll after which no further answers will be accepted (key: date)
-						* A boolean value that indicates the possibility of multiple answers from users (key: multiple-answers)
-						* A boolean value that indicates the possibility to edit the answers in future (key: allows-answer-edits)
-					2. --poll-answer 
-                       where you specify a slice of answers that will be provided to the users once they want to take part in poll votations.
-						Each answer should is identified by the text of the answer itself
-				Usage examples:
-				- tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "Post with poll" true \
-  					--poll-details "question=Which dog do you prefer?,multiple-answers=false,allows-answer-edits=true,end-date=2020-01-01T15:00:00.000Z" \
-  					--poll-answer "Beagle" \
-  					--poll-answer "Pug" \
-					--poll-answer "German Sheperd"
-		`),
+E.g.
+%s tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "Hello world!" true
+
+=== Medias ===
+If you want to add one or more media(s) attachment(s), you have to use the --media flag.
+You need to firstly specify the media URI and then its mime-type separeted by a comma.
+
+%s tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "A post with a single media" false \
+  --media "https://example.com/media1,text/plain"
+%s tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "A post with multiple medias" false \
+  --media "https://example.com/media1,text/plain"
+  --media "https://example.com/media2,application/json"
+
+
+=== Polls ===
+If you want to add a poll to your post you need to specify it through two flags:
+  1. --poll-details, which accepts a map with the following keys:
+     * question: the question of the poll
+     * date: the end date of your poll after which no further answers will be accepted
+     * multiple-answers: a boolean indicating the possibility of multiple answers from users
+     * allows-answers-edits: a boolean value that indicates the possibility to edit the answers in the future
+  2. --poll-answer, which accepts a slice of answers that will be provided to the users once they want to take part in the poll votations.	
+     Each answer should be identified by the text of the answer itself.
+
+E.g.
+%s tx posts create "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e" "Post with poll" true \
+	--poll-details "question=Which dog do you prefer?,multiple-answers=false,allows-answer-edits=true,end-date=2020-01-01T15:00:00.000Z" \
+	--poll-answer "Beagle" \
+	--poll-answer "Pug" \
+	--poll-answer "German Sheperd"
+`, version.ClientName, version.ClientName, version.ClientName, version.ClientName),
 		Args: cobra.MinimumNArgs(3),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			inBuf := bufio.NewReader(cmd.InOrStdin())
@@ -139,7 +139,10 @@ func GetCmdCreatePost(cdc *codec.Codec) *cobra.Command {
 			if len(pollDetailsMap) > 0 && len(pollAnswersSlice) > 0 {
 				date, err := time.Parse(time.RFC3339, pollDetailsMap[keyEndDate])
 				if err != nil {
-					return fmt.Errorf("end date should be provided in RFC3339 format, e.g 2020-01-01T12:00:00Z")
+					return fmt.Errorf(
+						"end date should be provided in RFC3339 format, e.g 2020-01-01T12:00:00Z, %s found",
+						pollDetailsMap[keyEndDate],
+					)
 				}
 
 				if date.Before(time.Now().UTC()) {
@@ -169,7 +172,7 @@ func GetCmdCreatePost(cdc *codec.Codec) *cobra.Command {
 					}
 
 					pollAnswer := types.PollAnswer{
-						ID:   uint(index),
+						ID:   types.AnswerID(index),
 						Text: answer,
 					}
 
@@ -308,14 +311,14 @@ func GetCmdAnswerPoll(cdc *codec.Codec) *cobra.Command {
 				return err
 			}
 
-			var answers []uint
+			var answers []types.AnswerID
 			for i := 1; i < len(args); i++ {
 				answer, err := strconv.ParseUint(args[i], 10, 32)
 				if err != nil {
 					return err
 				}
 
-				answers = append(answers, uint(answer))
+				answers = append(answers, types.AnswerID(answer))
 			}
 
 			msg := types.NewMsgAnswerPoll(postID, answers, cliCtx.FromAddress)