+++ x/posts/internal/keeper/handler_test.go
@@ -403,19 +403,19 @@ func Test_handleMsgRemovePostReaction(t *testing.T) {
 }
 
 func Test_handleMsgAnswerPollPost(t *testing.T) {
-	answers := []uint{uint(1), uint(2)}
-	userPollAnswers := types.NewAnswersDetails(answers, testPostOwner)
+	answers := []types.AnswerID{types.AnswerID(1), types.AnswerID(2)}
+	userPollAnswers := types.NewUserAnswer(answers, testPostOwner)
 
 	tests := []struct {
 		name          string
 		msg           types.MsgAnswerPoll
 		storedPost    types.Post
-		storedAnswers *types.AnswersDetails
+		storedAnswers *types.UserAnswer
 		expErr        error
 	}{
 		{
 			name: "Post not found",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(2), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, true, true),
@@ -424,14 +424,14 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "No poll associated with post",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil, nil),
 			expErr: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "no poll associated with ID: 1"),
 		},
 		{
 			name: "Answer after poll closure",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDateExpired, types.PollAnswers{answer}, true, false, true),
@@ -442,7 +442,7 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "Poll doesn't allow multiple answers",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer}, true, false, true),
@@ -452,7 +452,7 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "User provide too many answers",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2, 3}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2, 3}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, true, true),
@@ -462,7 +462,7 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "User provide answers that are not the ones provided by the poll",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 3}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 3}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, true, true),
@@ -472,7 +472,7 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "Poll doesn't allow answers' edits",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, true, false),
@@ -483,7 +483,7 @@ func Test_handleMsgAnswerPollPost(t *testing.T) {
 		},
 		{
 			name: "Answered correctly to post's poll",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testPostOwner),
+			msg:  types.NewMsgAnswerPoll(types.PostID(1), []types.AnswerID{1, 2}, testPostOwner),
 			storedPost: types.NewPost(types.PostID(1), types.PostID(0), "Post message", false, "desmos", map[string]string{}, testPostCreationDate,
 				testPostOwner, nil,
 				types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, true, true),