+++ x/posts/internal/types/genesis.go
@@ -2,9 +2,9 @@ package types
 
 // GenesisState contains the data of the genesis state for the posts module
 type GenesisState struct {
-	Posts       Posts                          `json:"posts"`
-	PollAnswers map[string]UsersAnswersDetails `json:"poll_answers_details"`
-	Reactions   map[string]Reactions           `json:"reactions"`
+	Posts       Posts                  `json:"posts"`
+	PollAnswers map[string]UserAnswers `json:"poll_answers_details"`
+	Reactions   map[string]Reactions   `json:"reactions"`
 }
 
 // NewGenesisState creates a new genesis state