+++ x/posts/internal/types/poll_answers_response_test.go
@@ -12,14 +12,14 @@ func TestPollAnswersQueryResponse_String(t *testing.T) {
 	testOwner, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
 	assert.NoError(t, err)
 
-	answers := types.AnswersDetails{
-		Answers: []uint{1, 2},
+	answers := types.UserAnswer{
+		Answers: []types.AnswerID{1, 2},
 		User:    testOwner,
 	}
 
 	pollResponse := types.PollAnswersQueryResponse{
 		PostID:         types.PostID(0),
-		AnswersDetails: types.UsersAnswersDetails{answers},
+		AnswersDetails: types.UserAnswers{answers},
 	}
 
 	assert.Equal(t, "Post ID [0] - Answers Details:\nUser: cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns \nAnswers IDs: 1 2", pollResponse.String())