+++ x/posts/internal/types/post_response_test.go
@@ -35,7 +35,7 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 	testPostEndPollDate := time.Date(2050, 1, 1, 15, 15, 00, 000, timeZone)
 	pollData := types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, false, true)
 
-	answers2 := []uint{uint(1)}
+	answers2 := []types.AnswerID{types.AnswerID(1)}
 
 	post := types.NewPost(
 		types.PostID(10),
@@ -75,7 +75,7 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		nil,
 	)
 
-	answersDetails := []types.AnswersDetails{types.NewAnswersDetails(answers2, liker)}
+	answersDetails := []types.UserAnswer{types.NewUserAnswer(answers2, liker)}
 
 	likes := types.Reactions{
 		types.NewReaction("like", liker),
@@ -93,7 +93,7 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		{
 			name:        "Post Query Response with Post that contains media and poll",
 			response:    PostResponse,
-			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","medias":[{"uri":"https://uri.com","mime_Type":"text/plain"}],"poll_data":{"question":"poll?","provided_answers":[{"id":1,"text":"Yes"},{"id":2,"text":"No"}],"end_date":"2050-01-01T15:15:00Z","is_open":true,"allows_multiple_answers":false,"allows_answer_edits":true},"poll_answers":[{"answers":[1],"user":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4"}],"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
+			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","medias":[{"uri":"https://uri.com","mime_Type":"text/plain"}],"poll_data":{"question":"poll?","provided_answers":[{"id":"1","text":"Yes"},{"id":"2","text":"No"}],"end_date":"2050-01-01T15:15:00Z","is_open":true,"allows_multiple_answers":false,"allows_answer_edits":true},"poll_answers":[{"answers":["1"],"user":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4"}],"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
 		},
 		{
 			name:        "Post Query with Post that not contains poll",
@@ -103,7 +103,7 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		{
 			name:        "Post Query Response with Post that not contains media",
 			response:    types.NewPostResponse(postNoMedia, answersDetails, likes, children),
-			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","poll_data":{"question":"poll?","provided_answers":[{"id":1,"text":"Yes"},{"id":2,"text":"No"}],"end_date":"2050-01-01T15:15:00Z","is_open":true,"allows_multiple_answers":false,"allows_answer_edits":true},"poll_answers":[{"answers":[1],"user":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4"}],"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
+			expResponse: `{"id":"10","parent_id":"0","message":"Post","created":"2020-02-02T15:00:00Z","last_edited":"0001-01-01T00:00:00Z","allows_comments":true,"subspace":"4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e","creator":"cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47","poll_data":{"question":"poll?","provided_answers":[{"id":"1","text":"Yes"},{"id":"2","text":"No"}],"end_date":"2050-01-01T15:15:00Z","is_open":true,"allows_multiple_answers":false,"allows_answer_edits":true},"poll_answers":[{"answers":["1"],"user":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4"}],"reactions":[{"owner":"cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4","value":"like"},{"owner":"cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae","value":"like"}],"children":["98","100"]}`,
 		},
 	}
 
@@ -112,10 +112,7 @@ func TestPostQueryResponse_MarshalJSON(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			jsonData, err := json.Marshal(&test.response)
 			assert.NoError(t, err)
-			assert.Equal(t,
-				test.expResponse,
-				string(jsonData),
-			)
+			assert.Equal(t, test.expResponse, string(jsonData))
 		})
 	}
 }
@@ -133,8 +130,8 @@ func TestPostQueryResponse_String(t *testing.T) {
 	timeZone, err := time.LoadLocation("UTC")
 	assert.NoError(t, err)
 
-	answers2 := []uint{uint(1)}
-	answersDetails := []types.AnswersDetails{types.NewAnswersDetails(answers2, liker)}
+	answers2 := []types.AnswerID{types.AnswerID(1)}
+	answersDetails := []types.UserAnswer{types.NewUserAnswer(answers2, liker)}
 
 	date := time.Date(2020, 2, 2, 15, 0, 0, 0, timeZone)
 	medias := types.PostMedias{
@@ -144,12 +141,12 @@ func TestPostQueryResponse_String(t *testing.T) {
 		},
 	}
 	answer := types.PollAnswer{
-		ID:   uint(1),
+		ID:   types.AnswerID(1),
 		Text: "Yes",
 	}
 
 	answer2 := types.PollAnswer{
-		ID:   uint(2),
+		ID:   types.AnswerID(2),
 		Text: "No",
 	}
 	pollData := types.NewPollData("poll?", time.Now().UTC().Add(time.Hour), types.PollAnswers{answer, answer2}, true, false, true)